+++ log/logger.go
@@ -114,6 +114,18 @@ func WithInstrumentationVersion(version string) LoggerOption {
 	})
 }
 
+// mergeSets returns the union of keys between a and b. Any duplicate keys will
+// use the value associated with b.
+func mergeSets(a, b attribute.Set) attribute.Set {
+	// NewMergeIterator uses the first value for any duplicates.
+	iter := attribute.NewMergeIterator(&b, &a)
+	merged := make([]attribute.KeyValue, 0, a.Len()+b.Len())
+	for iter.Next() {
+		merged = append(merged, iter.Attribute())
+	}
+	return attribute.NewSet(merged...)
+}
+
 // WithInstrumentationAttributes returns a [LoggerOption] that sets the
 // instrumentation attributes of a [Logger].
 //
@@ -121,15 +133,34 @@ func WithInstrumentationVersion(version string) LoggerOption {
 //
 // Note that [WithInstrumentationAttributeSet] is recommended as
 // it is more efficient and also allows safely reusing the passed argument.
+//
+// If multiple [WithInstrumentationAttributes] or [WithInstrumentationAttributeSet]
+// options are passed the attributes will be merged together in the order
+// they are passed. Attributes with duplicate keys will use the last value passed.
 func WithInstrumentationAttributes(attr ...attribute.KeyValue) LoggerOption {
+	if len(attr) == 0 {
+		return loggerOptionFunc(func(config LoggerConfig) LoggerConfig {
+			return config
+		})
+	}
+
 	return loggerOptionFunc(func(config LoggerConfig) LoggerConfig {
-		config.attrs = attribute.NewSet(attr...)
+		newAttrs := attribute.NewSet(attr...)
+		if config.attrs.Len() == 0 {
+			config.attrs = newAttrs
+		} else {
+			config.attrs = mergeSets(config.attrs, newAttrs)
+		}
 		return config
 	})
 }
 
 // WithInstrumentationAttributeSet returns a [LoggerOption] that sets the
 // instrumentation attributes of a [Logger].
+//
+// If multiple [WithInstrumentationAttributes] or [WithInstrumentationAttributeSet]
+// options are passed the attributes will be merged together in the order
+// they are passed. Attributes with duplicate keys will use the last value passed.
 func WithInstrumentationAttributeSet(set attribute.Set) LoggerOption {
 	return loggerOptionFunc(func(config LoggerConfig) LoggerConfig {
 		config.attrs = set